move analysis into solver
